requirements.txt
anyio==3.6.2
black==22.10.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
ecdsa==0.18.0
fastapi==0.88.0
filelock==3.8.2
flake8==6.0.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
identify==2.5.9
idna==3.4
inflect==6.0.2
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
pre-commit==2.20.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.1
websockets==10.4
